feat: 회원가입 단계별 이동 기능 활성화
